PLG-831: Move Key-indicator services
