+++ src/Akeneo/Pim/Automation/DataQualityInsights/tests/back/Specification/Application/KeyIndicator/GetKeyIndicatorsSpec.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Specification\Akeneo\Pim\Automation\DataQualityInsights\Application;
+namespace Specification\Akeneo\Pim\Automation\DataQualityInsights\Application\KeyIndicator;
 
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\Model\Read\KeyIndicator;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\Query\Dashboard\GetProductKeyIndicatorsQueryInterface;
